src/requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
AWSIoTPythonSDK==1.5.2
beautifulsoup4==4.10.0
blinker==1.4
boto3==1.26.121
botocore==1.29.121
Brlapi==0.8.3
Brotli==1.0.9
cached-property==1.5.2
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.7
distro==1.7.0
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
evdev==1.6.0
exceptiongroup==1.1.1
eyeD3==0.8.10
fastapi==0.87.0
filelock==3.12.3
filetype==1.2.0
flatbuffers==22.11.23
font-fredoka-one==0.0.4
frozenlist==1.3.3
glob2==0.7
greenlet==1.1.3.post0
grpcio==1.30.2
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
humanfriendly==10.0
idna==3.4
ifaddr==0.1.7
ImageHash==4.3.1
imageio==2.22.4
IMDbPY==2021.4.18
importlib-metadata==4.6.4
inky==1.5.0
jeepney==0.7.1
jmespath==1.0.1
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
llvmlite==0.39.1
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
more-itertools==8.10.0
mpmath==1.2.1
multidict==6.0.3
munch==2.5.0
mutagen==1.45.1
mysql-connector-python==9.1.0
nemo-emblems==6.0.1
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.8
numba==0.56.4
numpy==1.23.3
oauthlib==3.2.0
onboard==1.4.1
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
outcome==1.2.0
packaging==23.1
PAM==0.4.2
pexpect==4.8.0
Pillow==9.5.0
platformdirs==3.10.0
pooch==1.6.0
protobuf==3.12.4
psb==0.17.0
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pydantic==1.10.2
pyelftools==0.27
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyMatting==1.1.8
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-gnupg==0.4.8
python-magic==0.4.24
python-multipart==0.0.5
python-xlib==0.29
pytz==2022.1
PyWavelets==1.4.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
rembg==2.0.29
reportlab==3.6.8
requests==2.29.0
requests-file==1.5.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
RPi.GPIO==0.7.1
RuNAs==0.1
s3transfer==0.6.0
scikit-image==0.19.3
scipy==1.9.3
scour==0.38.2
SecretStorage==3.3.1
setproctitle==1.2.2
simplejson==3.17.6
six==1.16.0
smbus2==0.4.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
spidev==3.6
SQLAlchemy==2.0.36
sqlparse==0.4.4
ssh-import-id==5.11
starlette==0.21.0
sympy==1.11.1
systemd-python==234
texttable==1.6.4
tifffile==2022.10.10
tinycss2==1.1.1
tldextract==3.1.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.9.0
ubuntu-cleaner==1.1.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.15
uvicorn==0.20.0
virtualenv==20.24.4
wadllib==1.3.6
watchdog==2.1.9
webdriver-manager==4.0.0
webencodings==0.5.1
websocket-client==1.2.3
websockets==9.1
wget==3.2
wsproto==1.2.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
xlrd==1.2.0
yarl==1.8.2
youtube-dl @ git+https://github.com/ytdl-org/youtube-dl.git@8c86fd33dca48ebb505ed04150d9e35993b9fe7e
yt-dlp==2022.4.8
zipp==1.0.0
